ml-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
anthropic
transformers
torch
pillow
numpy
